backend/scraper/requirements.txt
asgiref==3.4.1
attrs==21.2.0
automat==20.2.0
beautifulsoup4==4.9.3
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4; python_version >= '3'
click==8.0.1
constantly==15.1.0
cryptography==3.4.7
cssselect==1.1.0
fastapi==0.66.0
h11==0.12.0
h2==3.2.0
hpack==3.0.0
httptools==0.2.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.2; python_version >= '3'
incremental==21.3.0
itemadapter==0.3.0
itemloaders==1.0.4
jmespath==0.10.0
lxml==4.6.3
parsel==1.6.0
priority==1.3.0
protego==0.1.16
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
pydispatcher==2.0.5; platform_python_implementation == 'CPython'
pymongo==3.12.0
pyopenssl==20.0.1
python-dotenv==0.19.0
python-multipart==0.0.5
pyyaml==5.4.1
queuelib==1.6.1
requests==2.26.0
scrapy==2.5.0
scrapyd==1.2.1
service-identity==21.1.0
six==1.16.0
soupsieve==2.2.1; python_version >= '3.0'
starlette==0.14.2
twisted==21.7.0
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.14.0
uvloop==0.15.3
w3lib==1.22.0
watchgod==0.7
websockets==9.1
zope.interface==5.4.0
